+++ src/soopervisor/abc.py
@@ -33,9 +33,9 @@ class AbstractConfig(BaseModel, abc.ABC):
         extra = 'forbid'
 
     @classmethod
-    def from_file_with_root_key(cls, path_to_config, env_name):
+    def from_file_with_root_key(cls, path_to_config, env_name, **defaults):
         # write defaults, if needed
-        cls._write_defaults(path_to_config, env_name)
+        cls._write_defaults(path_to_config, env_name, **defaults)
 
         data = yaml.safe_load(Path(path_to_config).read_text())
 
@@ -68,9 +68,11 @@ class AbstractConfig(BaseModel, abc.ABC):
         return cls(**data[env_name])
 
     @classmethod
-    def _write_defaults(cls, path_to_config, env_name):
-        data = cls.defaults()
-        default_data = yaml.safe_dump({env_name: data})
+    def _write_defaults(cls, path_to_config, env_name, **defaults):
+        data = {**cls.defaults(), **defaults}
+        # pass default_flow_style=None to it serializes lists as [a, b, c]
+        default_data = yaml.safe_dump({env_name: data},
+                                      default_flow_style=None)
 
         # if no config file, write one with env_name section and defaults
         if not Path(path_to_config).exists():
@@ -118,14 +120,6 @@ class AbstractExporter(abc.ABC):
         # ensure that the project and the config make sense
         self.validate()
 
-        # initialize configuration and a few checks on it
-        self._cfg = self.CONFIG_CLASS.from_file_with_root_key(
-            path_to_config=path_to_config,
-            env_name=env_name,
-        )
-
-        self._env_name = env_name
-
         # initialize dag (needed for validation)
         # TODO: _export also has to find_spec, maybe load it here and
         # pass it directly to _export?
@@ -133,6 +127,21 @@ class AbstractExporter(abc.ABC):
             spec, _ = commons.find_spec(cmdr=cmdr, name=env_name)
 
         self._dag = spec.to_dag().render(force=True, show_progress=False)
+
+        # it the spec has products store in relative paths, get them and
+        # exclude them
+        prod_prefix = commons.product_prefixes_from_spec(spec)
+        defaults = {} if not prod_prefix else dict(exclude=prod_prefix)
+
+        # initialize configuration (create file if needed) and a few checks on
+        # it
+        self._cfg = self.CONFIG_CLASS.from_file_with_root_key(
+            path_to_config=path_to_config,
+            env_name=env_name,
+            **defaults,
+        )
+
+        self._env_name = env_name
 
         # validate specific details about the target
         self._validate(self._cfg, self._dag, self._env_name)
@@ -145,6 +154,8 @@ class AbstractExporter(abc.ABC):
         commons.dependencies.check_lock_files_exist()
 
     def add(self):
+        """Create a directory with the env_name and add any necessary files
+        """
         # check that env_name folder does not exist
         path = Path(self._env_name)