+++ src/soopervisor/commons/__init__.py
@@ -1,5 +1,6 @@
 from soopervisor.commons import conda, docker, source, dependencies
-from soopervisor.commons.dag import load_tasks, find_spec
+from soopervisor.commons.dag import (load_tasks, find_spec,
+                                     product_prefixes_from_spec)
 
 __all__ = [
     'conda',
@@ -8,4 +9,5 @@ __all__ = [
     'load_tasks',
     'find_spec',
     'dependencies',
+    'product_prefixes_from_spec',
 ]