+++ src/soopervisor/exceptions.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 from click import ClickException
 
 
@@ -7,3 +9,16 @@ class ConfigurationError(ClickException):
     show the error message and not the whole traceback
     """
     pass
+
+
+class MissingDockerfileError(ClickException):
+    """
+    Raised when trying to build a Docker image but the DockerFile is missing
+    """
+    def __init__(self, env_name):
+        self.env_name = env_name
+        path = str(Path(env_name, 'Dockerfile'))
+        message = f"""\
+Expected Dockerfile at {path!r} but it does not exist\
+        """
+        super().__init__(message)