+++ tests/test_commons.py
@@ -2,6 +2,7 @@ import os
 import tarfile
 import subprocess
 from pathlib import Path
+from unittest.mock import Mock
 
 import yaml
 import pytest
@@ -12,6 +13,7 @@ from ploomber.io._commander import Commander
 
 from soopervisor.commons import source, conda, dependencies
 from soopervisor import commons
+from soopervisor.exceptions import MissingDockerfileError
 
 
 @pytest.fixture
@@ -266,16 +268,36 @@ def test_copy_ignore_git(tmp_empty):
 def test_compress_dir(tmp_empty):
     dir = Path('dist', 'project-name')
     dir.mkdir(parents=True)
-
     (dir / 'file').touch()
 
-    source.compress_dir('dist/project-name', 'dist/project-name.tar.gz')
+    with Commander() as cmdr:
+        source.compress_dir(cmdr, 'dist/project-name',
+                            'dist/project-name.tar.gz')
 
     with tarfile.open('dist/project-name.tar.gz', 'r:gz') as tar:
         tar.extractall('.')
 
     expected = {Path('project-name/file')}
     assert set(Path(p) for p in source.glob_all('project-name')) == expected
+
+
+def test_compress_warns_if_output_too_big(tmp_empty, monkeypatch, capsys):
+    # mock a file of 6MB
+    monkeypatch.setattr(source.os.path, 'getsize',
+                        Mock(return_value=1024 * 1024 * 6))
+
+    dir = Path('dist', 'project-name')
+    dir.mkdir(parents=True)
+    (dir / 'file').touch()
+
+    with Commander() as cmdr:
+        source.compress_dir(cmdr, 'dist/project-name',
+                            'dist/project-name.tar.gz')
+
+    captured = capsys.readouterr()
+    expected = ("The project's source code 'dist/project-name.tar.gz' "
+                "is larger than 5MB")
+    assert expected in captured.out
 
 
 @pytest.mark.parametrize('env_yaml, expected', [
@@ -425,3 +447,14 @@ def test_check_lock_files_exist(tmp_empty):
     expected = ('Expected requirements.lock.txt or environment.lock.yml at '
                 'the root directory')
     assert expected in str(excinfo.value)
+
+
+def test_error_if_missing_dockerfile(tmp_empty):
+    with pytest.raises(MissingDockerfileError) as excinfo:
+        commons.docker.build(e=Mock(),
+                             cfg=Mock(),
+                             name='some_name',
+                             until=Mock(),
+                             entry_point=Mock())
+
+    assert excinfo.value.env_name == 'some_name'